migration handler: Prevent adapter from continuing if migration handling fails
Currently, failures in the migration handlers will log their stack
trace, but the process will continue on, even though a core
functionality is broken. This diff will make the process abort in
the case of a task failure.

Follow-up diffs will focus on getting rid of the usage of unwraps
in the migration handler code, along with hopefully an improved
ability to gracefully recover from some failures.

Refs: ENG-1710
Change-Id: I96704fcf1fe81acabc76e934ae8896562807ca25
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2998
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>