test-utils: Add a proc macro to mark a test as slow
Add a simple `#[slow]` proc attribute macro to the test_utils, which
marks a function that should only be run if the `RUN_SLOW_TESTS`
environment variable is set.

This is (very slightly) more convenient than putting the conditional
directly in the test, but I hope enough of a difference that we might
actually start marking tests as slow.

Change-Id: I7c9b5b92f5253512a2e494bdeb489f67f9e82216
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3004
Tested-by: Buildkite CI
Reviewed-by: Nikolai Vazquez <nikolai@readyset.io>
Reviewed-by: Harley Klein <harley@readyset.io>