readyset-client: Refactor set handling to a single method
Refactor the way we classify set statements for the individual engines,
which currently is a pair of methods on the `QueryHandler` trait - one
to determine if a set is supported but we don't care about it, and one
to determine if a set is changing the autocommit state - into a single
method that returns an *enum* specifying how we should handle the given
set statement. This is being done in preparation for actually handling
other set statements directly ourselves in the way we currently do for
autocommit (specifically in my case this is for `SET search_path`).

Also, this commit starts to slightly modify the nomenclature here - now
instead of just saying a set statement is "supported", we specifically
call out that we are going to *proxy* that set statement in the naming
of the corresponding enum variant, to better differentiate between
"allowed, and we want to do something in response to this statement" and
"allowed, and we don't care and can just proxy upstream".

Refs: ENG-1584
Change-Id: I844539ffd7dcac730924b425f8de7ae410bb861c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3005
Reviewed-by: Peter Farr <peter@readyset.io>
Tested-by: Buildkite CI