readyset-psql: Allow changing search path
Allow changing the search_path in psql clients, by adding a new
`SetBehavior::SearchPath` variant, which is returned by the
`PostgreSqlQueryHandler::handle_set_statement` method for SET statements
that change the query path. Then, match on this new variant in
`readyset_client` to change the schema search path of the noria
connector in addition to proxying the statement.

Fixes: ENG-1584
Change-Id: Ic6ef6189e8eb2da37488f347e0da31cd2d507165
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3006
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>