readyset-mysql: Allow changing the current database with USE
Allow switching the currently active database with the `USE` statement,
in the same way that we currently allow switching the currently active
schema search path for postgresql.

Note that this is *not* currently validating that the db actually
exists; we don't track the set of databases that exist ourselves,
instead relying on the upstream to do that for us.

Refs: ENG-1583
Change-Id: Ibd705c27ed28dc6df5337b71422e96869cb31f98
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3011
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>