readyset-client: Thread schema_search_path through QueryStatusCache
It will soon become necessary for a correct schema_search_path to be
specified when installing a query. Furthermore, it will be possible for
two queries that're semantically equivalent to be semantically
different, if they're created with different schema search paths.

This commit doesn't *fully* fix that second problem, but does pave the
road to solving it - now, instead of just storing a SelectStatement, the
`QueryStatusCache` stores a new `PrepareRequest` struct (name pending),
which contains *all* the information needed to prepare a query,
which (at least for now) consists of the SelectStatement itself and the
schema search path. This is then provided to the migration_handler when
migrating the query asynchronously.

Refs: ENG-1737
Change-Id: Iff2f473f8a77d0504596e691232909dffa83524b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3013
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>