readyset-mysql: Add support for returning enum values in query results
This change allows us to get the actual string value of an enum column
correctly returned in a MySQL SELECT query result. In order to generate
the right string values, the SqlType values for the columns must be
passed into write_column so that the code that generates the query
response knows what the different enum elements are for a given column.

This change also saves the SqlType values in the schema cache. This is
partly for efficiency but mainly because it is very hard to implement
this change in any other way I've tried without upsetting the borrow
checker.

Note that there are still some rough edges on the underlying enum
support, and since this change will expose enums to clients I will try
to get some more CLs together and ready to merge in a batch so that we
have sufficient testing and don't ship this with known panic bugs.

Change-Id: I4a446fbe923809f8549005bb64905d0082d72728
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3020
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>