readyset-data: Eliminate extra single quotes in enum to string coercions
This fixes a bug where enum strings would show up surrounded by single
quote characters when they were returned from queries (for example we
would get back "'green'" instead of "green"). The issue was that the
Display impl for Literal would include single quotes to indicate the use
of the String variant. To fix this, we specifically match on
Literal::String and directly use its std::string::String value.

This should also cause us to error out if we somehow end up with a
non-String Literal in the list of elements for a SqlType::Enum, whereas
before we should have just blindly converted the value to a string.

Additionally, this should improve efficiency since we eliminate the
extra allocation associated with to_string(), and directly borrow the
underlying String instead.

Change-Id: I06f3f215d6ffde568c80eb2c5632c05256b6adbd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3031
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>