nom-sql: Rename Table to Relation, add some docs
Now that this struct is being used much more widely, calling it "Table"
is quite misleading - it's much better to use the more general name
"relation" for it, since that's really what it is - not to mention that
the name Table is also used for eg the RPC handler for tables in the
`readyset` crate. This commit renames the struct, and also updates a lot
of places where we were using it qualified to avoid naming conflicts to
not do so, since it no longer conflicts with `readyset::Table`.

Change-Id: I8a2dee0acfc53baf964515b1e0033df8b3add1fc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3032
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>