treewide: Remove "outputs" and "inputs" naming
Today there are (at least) 4 ways of referring to the thing that
`CREATE CACHE` creates:

 - "Views"
 - "Readers"
 - "Caches"
 - "Outputs"

And at least 3 ways of referring to the thing that `CREATE TABLE`
creates:

 - "Tables"
 - "Bases" / "Base Tables"
 - "Inputs"

The last element of each of those lists is by far the rarest, and (in my
opinion) the least intuitive, due to really being tied to the "dataflow
graph" metaphor for readyset. This commit removes both of those names
everywhere they're used in this way throughout the codebase, instead
opting to go with "Views" and "Tables" respectively, which seem to be
the most common and easy to follow.

Change-Id: If4be9aba0e145619c7c6f00c013fd2ab5c7d5a33
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3033
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>