controller: Add new RPC to request status of queries in bulk
Add a new RPC endpoint+method to the controller that takes a *list* of
`ViewCreateRequest`s (which include both a `SelectStatement` and a
schema search path) and returns a corresponding list (guaranteed to be
the same length) of booleans indicating whether or not those queries
have been installed in the graph (after rewriting them, which is
important). This is going to be used to replace the current behavior of
the `ViewsSynchronizer`, which doesn't currently function correctly with
queries that change semantics during their rewrite.

Refs: ENG-1737
Change-Id: I70bd1a573e82b692619b225d0631c10b2179a2bf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3036
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>