migration handler: Additionally handle Result::Err case
This compliments CL2998, except for Result::Err instead of panics.

When the migration code failed previously (panic/Result::Err), a
stack trace would be dumped to the console, and then the migration
task would go away, leaving the process in a semi-broken state.

These two make sure any failures that make it to the top of those
tasks prevent the process from continuing any further.

Refs: ENG-1710
Change-Id: Ib99b45dce4c643ed85144eb99e84b78330081666
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3037
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>