clustertest: Allow deploying multiple adapters
Add support to readyset-clustertest for deploying not just 0 or 1
but *any number* of adapter servers, and modify the various methods for
getting handles to those adapters to either just get the first
adapter (for the majority of tests that don't care which adapter they're
talking to) or specify which adapter to get a handle to. This will allow
testing the synchronization behavior between adapters, such as ensuring
that if one adapter creates a new query the ViewsSynchronizer in the
other adapters properly find out about that query.

Change-Id: Iaf1d927e84b9b147ce13bed5a01da65e19738dee
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3043
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>