metrics: make client metrics behave same as server
Currently the client only supports prometheus metrics
and installs them via the metrics crate. This change adds
support for noria-metrics in the client and installs them
using the install_global_recorder method, which makes the
metrics accessible to the http_router in standalone mode.

In addition makes the metrics installation code safe.

Change-Id: I9d3a7dd52c94b18f88341066a114a329a9ef674a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3046
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI