adapter: Provide server with a list of queries we care about
Now that the server is storing queries post-rewrite, our strategy of
loading a full list of queries from the server on an interval won't
work to synchronize explicitly migrated queries between multiple
adapters, because the adapter won't know about what the query will look
like post-rewrite if it's determining if it should execute it. Now,
instead, we use the newly introduced `view_statuses` API to provide the
controller with an explicit list of queries we care about (any queries
that are marked as pending in the query status cache), then use the
response to mark queries as explicitly migrated or not.

Refs: ENG-1737
Change-Id: Id3e9fb1e8ca5d64631e3500b0abc51374aa44ae0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3048
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>