readyset-client: Properly differentiate queries based on schema
Properly use the schema_search_path, by way of the `ViewCreateRequest`
field, to differentiate queries in the various structures inside of
readyset-client that cache things, most notably the ViewCache. This
allows two queries that have the same syntax, but a different meaning
due to being created with different schema search paths, to coexist
separately within the same adapter, and properly return different
results. This is proven with a new (originally failing) integration
test.

This isn't *all* that's necessary to fix this problem, however - if a
new table gets created that'd change the meaning of a previously-cached
query (eg because it skipped over that table's schema when looking for a
table with that name) the cache will still point to the wrong query.
Fixing that issue is up next.

Fixes: ENG-1737
Change-Id: If70beaf4989e2694332cc77bcb06fbca044beea6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3050
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>