readyset-mysql: resolve excessive buffering on fallback
When performing fallback queries, readyset-mysql would buffer the
entire resultset twice, once when buffering the result from
upstream, the second time when writing the downstream result into
a buffer. This changes now passes the result stream down to the
query handler instead, avoiding any buffering. All results are
written directly into the dowsntream and flushed periodically.

For very large resultsets the speedups are in excess of 10X.

The change required quite a bit of back and forth with the borrow
checker, and is rather large as a result.

readyset-psql should also benefit from similar optimizations.

Fixes: ENG-1728
Change-Id: I1a58f62f5444a1503989834f645e358ca1c9dcfa
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3074
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>