logictest: Add skipif/onlyif readyset
Allow supplying the string "readyset" to the skipif and onlyif logictest
conditionals to either skip or only run a command if we're running
against readyset (as opposed to running directly against mysql or
postgresql). This is generally a nice thing to have, but I'm doing it
now because I want to use it to send `CREATE SCHEMA` commands in a
logictest, but only if we're running against postgres (since readyset
doesn't care about a schema "existing" in order to create a table in
it).

Change-Id: I3d6e81643ee6f65a148c36cc10d73f5a6ef86bc8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3070
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>