sql-passes: Optionally output a list of "invalidating" tables
We want to make it so that the creation of a table that a
previously-existant query *would have* seen when resolving schemas
causes that query to be dropped - for examaple, if we have the
following:

    CREATE TABLE s2.t (x int);
    SET search_path = s1, s2;
    CREATE CACHE q FROM SELECT * FROM t;

because no table called `s1.t` exists, the query `q` would get rewritten
to:

    SELECT * FROM s2.t;

If a table called `s1.t` then gets *created*, that query is now
invalid - it *should* be `SELECT * FROM s1.t`. Rather than trying to
recreate the query, the plan here is to just drop it.

This commit implements the first step of that - now, an optional mutable
reference to a `Vec<Relation>` can be passed as part of the
`RewriteContext` to `readyset_sql_passes::Rewrite::rewrite` - if so, a
list of tables which if created should invalidate the query will be
pushed into that list. This is unit-tested, but nothing actually passes
that yet - next up, we'll thread that through and store it (likely in
`Recipe`) next to the query itself.

Refs: ENG-1747
Change-Id: Idd328dad3d80cf9dfaf86c8f543a091fa71d63d6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3071
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>