readyset-data: Remove extraneous src_type_name parameter
The coerce_integer function has this extra parameter which is only used
in error messages, and is currently hardcoded to specific strings in
each place where we call the function. Now that we're passing in both
from_ty and to_ty, all the info we need to generate good error messages
is now present in those types, so we can eliminate src_type_name
altogether.

Change-Id: I440e4c9552f3c6256cbb9c508e58f0579aaf7f5e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3075
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>