recipe: Rewrite registry tests to only cover the external API
Generally it's better style to have tests test the external API of a
unit rather than internal implementation details - that way, we
can *change* the internal implementation details without having to
change all of the tests. This rewrites the tests for the ExprRegistry to
only call exposed functions from the registry, rather than being tightly
coupled to the internal implementation.

Change-Id: Iaf16e0592d5f38f3d7a0aa9273e2d9a8c790799d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3077
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>