concrete-iter: Init proc macro for generating iter enums
Add a new crate concrete-iter, which provides a single attribute proc
macro for generating new concrete enum types for returning from a
function that returns `impl Iterator`. This is intended as a replacement
for either nested `itertools::Either` or `Box<dyn Iterator>`, without
the maintainability issues of the former or the performance impact of
the latter.

Change-Id: I9d72b9f5fe319ccf95ae1828cb732ff10783ee83
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3086
Tested-by: Buildkite CI
Reviewed-by: Nikolai Vazquez <nikolai@readyset.io>