clustertest: Extract start server/adapter params
This change is a step in aligning clustertests with the behavior of a
kubernetes deployment, which has separate configurations for a server
and adapter pod.

A DeploymentHandle now has ServerStartParams and AdapterStartParams that
are used when starting a new server (and in a future change, a new
adapter) in an ongoing deployment.

This means that some fields, for example the authority address, are now
in 2 places in a deployment instead of 1. This is intentional, as a
kubernetes deployment may be misconfigured such that an adapter and
server have different authority addresses. This will make the
clustertests more flexible in what failure modes can be tested.

This factor also includes some noria/readyset renaming.

Change-Id: I00d93a1d35db76325db1fd202dc64b3ab207fb87
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3085
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>