registry: Allow storing queries to invalidate for tables
Add a field to the ExprRegistry that allows storing a map from table
name to a set of queries to invalidate if that table ever gets created,
to allow queries to be invalidated if a table they should have seen when
resolving schemas gets created.

Nothing actually uses the methods for inserting into and accessing this
field yet; that's coming next.

Refs: ENG-1747
Change-Id: I62473676efe159d37a20ea1f8ff1272c2b67b1f7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3089
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>