recipe: Invalidate queries on creation of tables
If we've previously recorded that a query should be invalidated if a
table is created, actually do that invalidation upon the creation of
that table in the recipe, by calling `Self::remove_expression`.

Getting the `table_scope.test` logictest to pass as a result of this
requires letting the query error out one more time to bust the view
cache in the adapter - in the real world, this would just go to
fallback, which is fine since that'd return the right results.

Fixes: ENG-1747
Change-Id: Ifba3c88700e15b4135907d919e1e14406251ec57
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3094
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>