logictests: Add psql logictests
Add a separate directory for logictests that need to be run with
`--database-type postgresql`, so that we can test postgresql-specific
extensions and commands, including (in my case) the search_path.

Since I'm using this for testing the search_path, I've also moved
`table_scope.test` under that directory, in preparation for adding some
postgresql-specific stuff to it next.

Change-Id: I1b7caaaf0a2da2d47c463f7629d5f665320567a7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3093
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>