logictests: Cover changing schema search path
Dan pointed out a potential issue with search path where we might not
end up getting entries in `ExprRegistry::table_to_invalidated_queries`
because the search path was changed, and the query retried, after the
query was originally created with a search path that wouldn't cause it
to be invalidated by a table. It turns out that's not an issue because
the adapters generate different names for queries with different search
paths, causing them to go through the full query creation flow including
the addition of invalidating tables - this passing addition to the
table_scope test demonstrates that.

Change-Id: I365cf2e5c50d1e6586ccd9cfa8ad6d512ea7937d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3103
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>