Initial Health Reporter Design
This submits a basic design for a shared health reporter. The health
reporter allows a consumer to set the current state of health for the
adapter. Doing so will automatically update a state transition time if
the state has changed, and if the state is the same as the current
state, no write operation occurs.

Change-Id: I21e75f6167433302747f1dc8b421849fccbf5886
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3104
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>