clustertest: Add tracing support
This commit adds readyset-tracing and tracing as dependencies to
readyset-clustertests so that they can be used to instrument the
clustertest with helpful log messages.

It also extends the #[clustertest] macro to always the test logging and
print out the name of the test being run as a DEBUG log when the tests
start.

Change-Id: Ib36f96e1c50337d88d8c66b20cf2e64c797fdc7a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3110
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>