replicators: Allow specifying the snapshot reporting interval
This adds an option (--snapshot_report_interval_secs) to allow
specifying the interval between progress reports for snapshotting,
in seconds. If the value of 0 is specified, report is disabled. As
a u16 is used, the max time between reports is a bit over 18 hours
(18 hours ought to be enough for anybody?)

Additionally, because why not, it now also shows an estimate of the
remaining time to complete the snapshotting.

Refs: ENG-1660
Change-Id: I558fddedff8a307a8ac937ef7c2422c11db24f86
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3116
Tested-by: Buildkite CI
Reviewed-by: Alex Martin <Alex.m@readyset.io>