dataflow: Still update row count for groups if value doesn't change
cb163b8e0 (Keep count for aggregates in state, 2022-06-30) fixed an
issue where if we deleted the last row for a group, grouped operators
wouldn't remove that group from their output, by making every grouped
operator keep track of the number of rows per group in its own state,
and completely revoking the group if that number of rows ever dropped to
zero. This tracking had an edge condition where if a batch of deltas
ever left an aggregate value completely unchanged (eg if we added 0 to a
SUM), we wouldn't actually update the row count, meaning we wouldn't
actually revoke the group when removing all the rows. This fixes that by
swapping a comparison inside of grouped/mod.rs to check if the total row
count in the group hasn't changed *from the original value*, rather
than just checking that it's greater than zero, to determine if we need
to emit diffs.

Fixes: ENG-1762
Change-Id: Ib2c2c4ac285c0eb89e1614404258550f1099f76f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3122
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>