readyset-dataflow: Coerce enum values before inserting in base nodes
Prior to this change, we had an issue with the replicator where snapshot
replication would insert enum values as strings instead of as their
underlying integer values. Rather than address this at the level of the
replicator, Griffin suggested that it should be impossible to insert a
value of the wrong type into a base table, and that we should add code
to the base node to coerce values to their correct types according to
the underlying schema.

While it would be nice to have all types checked before we insert them
into a base node, this change just checks enum values since it's the
only known issue like this right now, and it would be a lot of work to
write new code to check and coerce all existing types. However the
addition of this code does lay groundwork that would make it easier to
add more of these kinds of checks in the future.

Note that there are no tests included in this commit because there is a
separate commit coming after this one with a new replicator test that is
able to reproduce the enum problem described above. With this commit in
place the test passes.

Change-Id: I7e212c8c3866b82e4d00fd4ddbf1360878e797bd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3126
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>