test: Add replication test for MySQL enums
This test initially failed because snapshot replication would retrieve
the string values for enums instead of the integer values (binlog
replication does not seem to have this problem, however). Prior to
committing this test, I wrote and committed a change to the base node
code to ensure that enum values are converted to integers prior to being
inserted into base tables, so this test should now pass.

Change-Id: I0904796830443393a5e2003051b81aba0cb02663
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3127
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>