Plumb health reporter through adapter
This commit adds basic plumbing of the adapters health reporter through
to its http router, to be used to return a status code to the health
endpoint.

Currently it just models the same behavior we currently do of returning
a 200, since the health reporter is initialized with the Initializing
state.

In a quick follow CL I will fill in health reporting throughout the
adapter, and add logic for reporting unhealthy if we have been
initializing for too long.

Change-Id: I387acc939eacbb9da1147da37192b7ed22d30cad
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3133
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>