readyset-data: Add code to handle text lengths in enum coercions
Prior to this change, coercing an enum type to CHAR(N) or VARCHAR(N) did
not respect the lengths specified in those types. With this new code in
place, coercing to a longer CHAR type will pad the result with
whitespace, and coercing to a shorter CHAR/VARCHAR type will truncate
the resulting string accordingly.

Note that this logic is somewhat copied from text::coerce_to, which I
would normally not want to do, since I'd prefer not to have this code
duplicated more than necessary. But, I know that Nikolai is working on a
function to handle this type of thing the same way everywhere, so for
now I'm fine with copying in a quick fix since I expect this code to be
replaced at some point in the not-too-distant future.

This commit also adds corresponding unit test cases. Without the new
code in coerce_enum, some of these test cases fail, but they should all
pass now.

Change-Id: I945a46b38eb9d0000fef07e4d81fa4a4bdd526d7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3144
Tested-by: Buildkite CI
Reviewed-by: Nikolai Vazquez <nikolai@readyset.io>