dataflow: fix duplicate replays for range queries
When a reader replay is triggered for a range key, it is first
checked if the interval is covered by the map. This is usually
fine, since replays only triggered for misses. However if there
is a race between the moment the trigger is processed and another
partial range is filled in the map, the check is invalid. The fix
is to only trigger replay for parts of the range that are still
missing in the map, and remove ranges that have been filled.

Fixes: ENG-1624
Change-Id: I32d262f9f66cf767fa56df66baced4083ecc3054
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3148
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>