telemetry: Make telemetry reporting async
This change Introduces TelemetrySender, which is a wrapper around a
tokio mpsc channel and forwards metric send events to the
TelemetryReporter, which has been changed to be a background task.

Startup behavior between the installer/server/adapter is consolidated
with TelemetryInitializer::init() introduced in this crate.

The behavior of the telemetry reporter should be the same, with the
difference that we no longer block waiting for metric events to be sent
to segment.

Change-Id: I254e557edc986987f67e1ac081ff1a41ca016d15
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3150
Tested-by: Buildkite CI
Reviewed-by: Alex Martin <Alex.m@readyset.io>