telemetry: convert panics to errors
With this change, processes that use the telemetry reporter (i.e.
readyset-server, readyset-adapters, and the installer) will not be
affected by any panics in the Telemetry initialization/reporting, as
they have been converted to errors.

Users of the telemetry reporter can decide whether to keep running upon
encountering an error or propagate up the errors. The current behavior
is to keep running for all 3, with the reasoning that we don't want a
non-user visible feature to impact user-visible ones.

Change-Id: I4f43f7cccbe2b4b035200477c99ee19a4873535f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3152
Tested-by: Buildkite CI
Reviewed-by: Alex Martin <Alex.m@readyset.io>