nom-sql: Support MySQL-specific CAST syntax for int types
The CAST function in MySQL is a bit odd, because it does not support all
MySQL types; instead it supports a limited subset of them, and
additionally, CASTing to integer types requires using special
CAST-specific keywords.

To be more precise, out of the various integer types it supports, MySQL
only allows casting to BIGINT types, and to do so, you must write
`CAST(i AS SIGNED [INTEGER])` or `CAST(i as UNSIGNED [INTEGER])`.

Currently, the ReadySet CAST function does not implement any of these
quirks, and instead behaves more like a normal ANSI SQL CAST in that it
simply accepts any data type. This commit does *not* restrict the
existing behavior to match the MySQL limitations, but it *does* add
support for the aliases that MySQL uses for casting to integer types
(i.e. when parsing with the MySQL dialect, this commit enables
expressions like `CAST(i AS UNSIGNED)` that ReadySet did not previously
support).

Fixes: ENG-708
Change-Id: Ifed08e45bcdc2debe72efb3d218949320db82f26
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3162
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>