Add telemetry field deployment_env
This change adds a new field to be populated on all telemetry messages,
identifying the environment in which corresponding ReadySet instance is
deployed, e.g. docker-compose via the ReadySet orchestrator, Helm via
the official ReadySet Helm chart, etc.

Currently, it can have one of the following values:
- InstallerCompose
- Eks
- Helm

The deployment_env field is set using the DEPLOYMENT_ENV environment
variable; ReadySet binaries do not otherwise attempt to deduce the
environment they are running in.

Refs: ENG-1714
Change-Id: I4258e021506744a7cdd0ce22bf5b67e29a7fcd40
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3153
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>