dataflow: Rewrite evictions through straddled joins
eca3fec38 (dataflow: Rewrite evictions through generated columns,
2022-07-19) solved the issue where we drop evictions if nodes generate
columns *from themselves* (eg paginate), but didn't solve the equivalent
issue for nodes which generate columns *from their parents* (eg
straddled joins). Essentially, the approach in eca3fec38 was that we'd
sift through the keys being evicted, remapping the columns in the
upstream keys to columns in the downstream keys directly. On one hand,
this doesn't work for joins since the columns being evicted from are
columns in the parent node, not columns in the current node; but perhaps
more significantly this is an inherently incorrect approach because the
design of key remapping allows nodes to remap downstream keys to /any
upstream key/, rather than just being limited to only selecting specific
columns (which is coincidentally what we currently only do, but doesn't
necessarily have to be, eg `project` could remap queries like `where x +
1 = ?` by performing the inverse operation to the upstream key).

Instead of that approach, this commit stores a *mapping* in the domain
from upstream keys to downstream keys which have remapped into that
upstream key; we then query that mapping as part of
`ReplayPaths::downstream_dependent_paths` to remap keys when handling
evictions through remapped keys. This makes evictions process correctly
through straddled joins, plus lays the groundwork to removing the
approach added in eca3fec38 (though that approach is still present as of
this change).

This includes a logictest to test this behavior, which uses the usual
trick of an evicty-join to reliably trigger evictions in the graph.

Fixes: ENG-1572
Change-Id: Iaa37797bdf2ec0f2a6a52ab4f27de642d620dd21
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3178
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>