Move anonymize visitor to readyset-sql-passes
This hoists the anonymize_literal function and associated visitors from
readyset-client/rewrite.rs to readyset-sql-passes/anonymize.rs, which is
a more reasonable dependency to add for other crates that may need to do
an anonymization pass (like we will need to do to anonymize schemas for
segment).

The existing tests for anonymize_literals consisted of processing a
query then anonymizing it and were split into separate tests for the
processing and anonymization.

Change-Id: Ie4c739bc0f2c44d8c6adbe9c5741e1c93221d155
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3195
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>