client-test-helpers: Refactor test env setup to a Builder
The various `setup`-esque functions for setting up test adapters in
`readyset-client-test-helpers` have *long* passed the point where adding
new arguments to them is sustainable (they did as soon as we had three
adjacent boolean arguments). I want to add even more configuration to
the setup of adapter tests, so in preparation this commit refactors all
of these functions to instead be just invocations of a new `TestBuilder`
struct, which makes it much more obvious what's going on at the
callsite, and paves the way for doing even more interesting things with
configuration.

Change-Id: I4c6b6fe407afd538679603e83576d474bc025821
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3199
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>