nom-sql: Make ColumnConstraint::DefaultValue take an Expr
Both MySQL and PostgreSQL accept any expression as the DEFAULT of a
column, not just a literal - plus having Literal here was the reason we
made CURRENT_TIMESTAMP etc. variants of Literal despite not actually
being literals. This just makes it an Expr, which is relatively
straightforward to do given that we don't actually need to handle
default values for columns ourselves (there's some stuff in mir_to_flow
that threads literal default values to base nodes in dataflow, but
nothing actually depends on that).

Since function exprs take precedence over literals, this slightly
changes the AST output of parsing eg `current_timestamp()` - this is
intentional, and paves the way to actually removing things like
`Literal::CurrentTimestamp`.

Refs: ENG-1829
Change-Id: Ib0be7a116169e0973cc4c3d232e25536e873c945
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3202
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>