nom-sql: Make Insert data Exprs
Similar to the DEFAULT value for columns, both MySQL and PostgreSQL
support arbitrary expressions in the data for INSERT statements, not
just literals. We only actually need to handle insert statements
ourselves in tests (otherwise everything just goes to upstream) so we
can get away with just returning an unsupported error for non-literal
values here - but making this Expr allows us to later remove
Literal::CurrentTimestamp and friends

Since this now makes eg `-1` in logictests parse as a unary negation of
a literal `1` rather than a standalone literal -1, this also requires an
ad-hoc implementation of unary negation in the handling of inserting in
readyset-client - this is just for tests, and should not be extended
further.

Refs: ENG-1829
Change-Id: Iae0a22914d3fae95a3f5c55d0482b2caac620c60
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3203
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>