adapters: Run integration tests in parallel
Up to this point, we've been limited to always running integration tests
for adapters which depend on an upstream database in serial, because
otherwise the tests would connect to the same database and end up
stepping on each others' toes. To allow our tests to run in parallel, we
can instead generate a unique database name for each test (and
server_id, in the case of replication from mysql) and recreate that
database at the beginning of each test case. This removes the
requirement for tests that use readyset_client_test_helpers and depend
on an upstream database to run serially, which should *significantly*
speed up running these tests in CI.

Getting this working required upping the `max_wal_senders` flag for the
postgres server we run for tests, since otherwise we run into the
maximum number of replication slots pretty quickly.

Change-Id: Ia9ecfc1a6daed39f54153014eceb934548fc8dc6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3205
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>