concrete-iter: Support iterators bound by lifetimes
The main trick here is to sniff out all the named lifetime parameters in
the item type, and then list those as lifetimes in the `impl Iterator`
for the returned iterator. This obviously doesn't work for elided
lifetimes yet (since there aren't actually any named lifetime
parameters!) but that can be fixed by manually de-eliding lifetimes
ourselves if we want (later).

Change-Id: I74ef180bffde0d574a09b10122f73fe71cf658b5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3212
Tested-by: Buildkite CI
Reviewed-by: Nikolai Vazquez <nikolai@readyset.io>